refactor(server): extract publishing-channels routes (2/3 of publishing)
Second of the 3-way publishing split. 4 channel routes (channels list,
channels POST, channels/reddit/allowed-subreddits, channels/:id DELETE)
-> src/server/routes/publishing-channels.js, mounted
`app.use('/api/publishing', publishingChannelsRouter)` (shares the
/api/publishing prefix with publishing-queue + the upcoming dispatcher).

Auth: all 4 are requireAuth, but kept PER-ROUTE (not mount-level) — a
mount-level requireAuth on a shared-prefix router would leak onto the
other /api/publishing/* routes that fall through to this mount.

Deps: pool, requireAuth/verifyBrandAccess (auth). validateSubName (reddit
sub-name validator) is an inner closure -> moved with its handler. No
blockers.

Test plan:
- node --check server.js + module -> OK
- npm run lint (no-undef) -> clean (first pass)
- module boot-load (import()) -> resolves
- route-inventory guard -> 211 byte-identical (20 publishing paths: 14
  queue + 4 channels reconstructed + 2 publish still inline)
- vitest -> 93/93

Rollback: revert; verbatim move.
